+++ Globus/CoreGraphics/GLBPoint.m
@@ -65,8 +65,8 @@ CGPoint GLBPointRotateAroundPoint(CGPoint point, CGPoint pivot, CGFloat angle) {
 }
 
 CGFloat GLBPointDistance(CGPoint p1, CGPoint p2) {
-    CGFloat pow1 = (CGFloat)GLB_POW(p1.x - p2.x, 2);
-    CGFloat pow2 = (CGFloat)GLB_POW(p1.y - p2.y, 2);
+    CGFloat pow1 = GLB_POW(p1.x - p2.x, 2);
+    CGFloat pow2 = GLB_POW(p1.y - p2.y, 2);
     return GLB_SQRT(pow1 + pow2);
 }
